services/backend/requirements.txt
Authlib==1.2.1
bcrypt==4.0.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.3
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
google-api-core==2.11.0
google-api-python-client==2.87.0
google-auth==2.19.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
httplib2==0.22.0
idna==3.4
importlib-metadata==6.5.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
oauth2client==4.1.3
oauthlib==3.2.2
Pillow==9.5.0
protobuf==4.23.2
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.0.9
PySenseSDK @ git+https://github.com/alexfrisbie/PySense.git@3bfa91415a189dcb3a7cee3d21331e00626d2847
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.10
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
